refactor: move PBC code into own file
