+++ src/kd_tree/query/mod.rs
@@ -1,6 +1,7 @@
 mod approx_nearest_one;
 mod best_n_within;
 mod builder;
+mod periodic;
 pub use builder::{
     ApproxNearestOneQuery, BestNWithinQuery, Exclude, Include, NearestNQuery,
     NearestNUnsortedQuery, NearestNWithinQuery, NearestNWithinUnsortedQuery, NearestOneQuery,
